# api/app_management/openapi.yaml
openapi: 3.0.3
info:
  title: CasaOS App Management API
  version: v2
  description: |
    <picture>
        <source media="(prefers-color-scheme: dark)" srcset="https://raw.githubusercontent.com/IceWhaleTech/logo/main/casaos/casaos_banner_dark_night_800px.png">
        <source media="(prefers-color-scheme: light)" srcset="https://raw.githubusercontent.com/IceWhaleTech/logo/main/casaos/casaos_banner_twilight_blue_800px.png">
        <img alt="CasaOS" src="https://raw.githubusercontent.com/IceWhaleTech/logo/main/casaos/casaos_banner_twilight_blue_800px.png">
    </picture>

    CasaOS App Management manages app lifecycle, including installing, uninstalling, starting, stopping, setting, etc.

    For issues and discussions, please visit the [GitHub repository](https://github.com/IceWhaleTech/CasaOS) or join [our Discord](https://discord.gg/knqAbbBbeX).
servers:
  - url: /v2/app_management
...
security:
  - access_token: []
paths:
...
  /apps:
    get:
      summary: Get app list from AppStore
      description: |
        Get lists of StoreInfo of all compose apps from AppStore.

        - `recommend` - Compose apps recommended by CasaOS team
        - `list` - All compose apps
        - `community` - Compose apps contributed by community
      operationId: composeAppStoreInfoList
      tags:
        - AppStore methods

      responses:
        "200":
          $ref: "#/components/responses/ComposeAppStoreInfoListsOK"
        "500":
          $ref: "#/components/responses/ResponseInternalServerError"
  /apps/{id}:
    get:
      summary: Get app info from AppStore
      description: |
        Get StoreInfo of specific compose app.

        > A compose app contains one or more container `apps`. One of the `apps` is the `main_app` of this compose app, whereas others are dependency `apps`.
      operationId: composeAppStoreInfo
      tags:
        - AppStore methods
      parameters:
        - $ref: "#/components/parameters/AppStoreID"
      responses:
        "200":
          $ref: "#/components/responses/ComposeAppStoreInfoOK"
        "404":
          $ref: "#/components/responses/ResponseNotFound"
        "500":
          $ref: "#/components/responses/ResponseInternalServerError"
  /apps/{id}/compose:
    get:
      summary: Get compose details of an app from AppStore
      description: |
        Accept both `application/json` and `application/yaml` content type for the `Accept` header in the HTTP request.

        If the `Accept` header is set to `application/yaml`, the response will be the original `docker-compose.yml` content that follows [Compose Specification](https://compose-spec.io).
      operationId: composeApp
      tags:
        - AppStore methods
      parameters:
        - $ref: "#/components/parameters/AppStoreID"
      responses:
        "200":
          $ref: "#/components/responses/ComposeAppOK"
        "404":
          $ref: "#/components/responses/ResponseNotFound"
        "500":
          $ref: "#/components/responses/ResponseInternalServerError"
  /compose:
    get:
      summary: Get the list of installed compose apps
      description: |
        TODO
      operationId: myComposeAppList
      tags:
        - Compose methods
      responses:
        "200":
          $ref: "#/components/responses/ComposeAppListOK"
        "500":
          $ref: "#/components/responses/ResponseInternalServerError"

    post:
      summary: Install a compose app
      description: |
        TODO
      operationId: installComposeApp
      tags:
        - Compose methods
      requestBody:
        $ref: "#/components/requestBodies/RequestComposeApp"
      responses:
        "200":
          $ref: "#/components/responses/ComposeAppInstallOK"
        "400":
          $ref: "#/components/responses/ResponseBadRequest"
        "500":
          $ref: "#/components/responses/ResponseInternalServerError"
  /compose/{id}:
    get:
      summary: Get compose details of a locally installed app
      description: |
        Accept both `application/json` and `application/yaml` content type for the `Accept` header in the HTTP request.

        If the `Accept` header is set to `application/yaml`, the response will be the interpolated `docker-compose.yml` content that follows [Compose Specification](https://compose-spec.io).

        > The difference between this and `/apps/{id}/compose` is that the compose detail returned has all environment variables interpolated. E.g. `$TZ` will be replaced by actual timezone text.
      operationId: myComposeApp
      tags:
        - Compose methods
      parameters:
        - $ref: "#/components/parameters/ComposeAppID"
      responses:
        "200":
          $ref: "#/components/responses/ComposeAppOK"
        "404":
          $ref: "#/components/responses/ResponseNotFound"
        "500":
          $ref: "#/components/responses/ResponseInternalServerError"

    put:
      summary: Update settings of a compose app
      description: |
        TODO
      operationId: updateComposeAppSettings
      tags:
        - Compose methods
      parameters:
        - $ref: "#/components/parameters/ComposeAppID"
      requestBody:
        $ref: "#/components/requestBodies/RequestComposeApp"
      responses:
        "200":
          $ref: "#/components/responses/ComposeAppUpdateSettingsOK"
        "400":
          $ref: "#/components/responses/ResponseBadRequest"
        "404":
          $ref: "#/components/responses/ResponseNotFound"
        "500":
          $ref: "#/components/responses/ResponseInternalServerError"
  /compose/{id}/status:
    get:
      summary: Get the status of a compose app
      description: |
        TODO
      operationId: composeAppStatus
      tags:
        - Compose methods
      parameters:
        - $ref: "#/components/parameters/ComposeAppID"
      responses:
        "200":
          $ref: "#/components/responses/ComposeAppStatusOK"
        "404":
          $ref: "#/components/responses/ResponseNotFound"
        "500":
          $ref: "#/components/responses/ResponseInternalServerError"
  /compose/{id}/containers:
    get:
      summary: Get the list of containers of a compose app
      description: |
        TODO
      operationId: composeAppContainers
      tags:
        - Compose methods
      parameters:
        - $ref: "#/components/parameters/ComposeAppID"
      responses:
        "200":
          $ref: "#/components/responses/ComposeAppContainersOK"
        "404":
          $ref: "#/components/responses/ResponseNotFound"
        "500":
          $ref: "#/components/responses/ResponseInternalServerError"
  /compose/{id}/logs:
    get:
      summary: Get the logs of a compose app
      description: |
        TODO
      operationId: composeAppLogs
      tags:
        - Compose methods
      parameters:
        - $ref: "#/components/parameters/ComposeAppID"
        - name: lines
          in: query
          description: |
            Number of lines to show from the end of the logs for each compose app.

            - show all logs if `-1`
          schema:
            type: integer
            default: 1000
      responses:
        "200":
          $ref: "#/components/responses/ComposeAppLogsOK"
        "404":
          $ref: "#/components/responses/ResponseNotFound"
        "500":
          $ref: "#/components/responses/ResponseInternalServerError"
...
components:
  securitySchemes:
    access_token:
      type: apiKey
      in: header
      name: Authorization
  parameters:
    AppStoreID:
      name: id
      description: App Store ID
      in: path
      required: true
      schema:
        type: string
        example: "e4380876-d759-4b26-af1f-968a4606d399"
    ComposeAppID:
      name: id
      description: App ID of the compose app (equivalent to compose project name)
      in: path
      required: true
      schema:
        type: string
        example: "syncthing-jsaq"
...
  requestBodies:
    RequestComposeApp:
      description: Compose app YAML content
      required: true
      content:
        application/yaml:
          schema:
            $ref: "#/components/schemas/ComposeApp"
          example: |
            # docker compose YAML content
            services:
              ...
        application/json:
          schema:
            $ref: "#/components/schemas/ComposeApp"
  responses:
    ResponseInternalServerError:
      description: Internal Server Error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/BaseResponse"
          example:
            message: "Internal Server Error"
    ResponseNotFound:
      description: Not Found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/BaseResponse"
          example:
            message: "Not Found"
    ResponseBadRequest:
      description: Bad Request
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/BaseResponse"
          example:
            message: "Bad Request"
...
    ComposeAppStoreInfoListsOK:
      description: OK
      content:
        application/json:
          schema:
            allOf:
              - $ref: "#/components/schemas/BaseResponse"
              - properties:
                  data:
                    $ref: "#/components/schemas/ComposeAppStoreInfoLists"
    ComposeAppStoreInfoOK:
      description: OK
      content:
        application/json:
          schema:
            allOf:
              - $ref: "#/components/schemas/BaseResponse"
              - properties:
                  data:
                    $ref: "#/components/schemas/ComposeAppStoreInfo"
    ComposeAppListOK:
      description: OK
      content:
        application/json:
          schema:
            allOf:
              - $ref: "#/components/schemas/BaseResponse"
              - properties:
                  data:
                    type: object
                    additionalProperties:
                      $ref: "#/components/schemas/ComposeAppWithStoreInfo"
                    example:
                      syncthing: {}
    ComposeAppInstallOK:
      description: OK
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/BaseResponse"
    ComposeAppOK:
      description: OK
      content:
        application/json:
          schema:
            allOf:
              - $ref: "#/components/schemas/BaseResponse"
              - properties:
                  data:
                    $ref: "#/components/schemas/ComposeAppWithStoreInfo"
        application/yaml:
          schema:
            type: string
    ComposeAppUpdateSettingsOK:
      description: OK
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/BaseResponse"
    ComposeAppStatusOK:
      description: OK
      content:
        application/json:
          schema:
            allOf:
              - $ref: "#/components/schemas/BaseResponse"
              - properties:
                  data:
                    type: string
                    example: "running"
    ComposeAppContainersOK:
      description: OK
      content:
        application/json:
          schema:
            allOf:
              - $ref: "#/components/schemas/BaseResponse"
              - properties:
                  data:
                    $ref: "#/components/schemas/ComposeAppContainers"
    ComposeAppLogsOK:
      description: OK
      content:
        application/json:
          schema:
            allOf:
              - $ref: "#/components/schemas/BaseResponse"
              - properties:
                  data:
                    readOnly: true
                    type: string
                    example: <logs...>
...
  schemas:
    BaseResponse:
      properties:
        message:
          readOnly: true
          description: message returned by server side if there is any
          type: string
          example: ""
...
    ComposeAppStoreInfoLists:
      properties:
        recommend:
          type: object
          additionalProperties:
            $ref: "#/components/schemas/ComposeAppStoreInfo"
        list:
          type: object
          additionalProperties:
            $ref: "#/components/schemas/ComposeAppStoreInfo"
        community:
          type: object
          additionalProperties:
            $ref: "#/components/schemas/ComposeAppStoreInfo"
    ComposeAppStoreInfo:
      properties:
        appstore_id:
          title: StoreID of the compose app
          description: |
            The StoreID is the unique identifier of the app in the AppStore.
          readOnly: true
          type: string
          example: "syncthing.syncthing (developer + app name)" # TODO
          x-go-name: AppStoreID
          x-oapi-codegen-extra-tags:
            mapstructure: "appstore_id"
            yaml: "appstore_id,omitempty"
        main_app:
          title: name of the main container app in `apps`
          description: |
            The main container app is one of the container `apps` that the user will interact with.
          type: string
          example: "syncthing"
          x-oapi-codegen-extra-tags:
            mapstructure: "main_app"
            yaml: "main_app,omitempty"
        apps:
          title: container apps included in the compose app
          description: |
            The key is the name of a container app, and the value is the StoreInfo of this container app.
          readOnly: true
          type: object
          additionalProperties:
            $ref: "#/components/schemas/AppStoreInfo"
          example:
            syncthing: {}
          x-oapi-codegen-extra-tags:
            yaml: ",omitempty"
    ComposeApp:
      type: object
      description: |-
        See [Compose Specification](https://compose-spec.io) for the schema structure of `ComposeApp`.

      x-go-type: types.Project
      x-go-type-import:
        name: types
        path: github.com/compose-spec/compose-go/types
    ComposeAppWithStoreInfo:
      description: |-
        Schema for `ComposeAppOK` response.

        > Do not mistake it as `ComposeAppStoreInfo` which is just store information of a compose app.
      properties:
        store_info:
          $ref: "#/components/schemas/ComposeAppStoreInfo"
        compose:
          $ref: "#/components/schemas/ComposeApp"
    AppStoreInfo:
      title: StoreInfo of an app
      description: |-
        > Do not mistake it as *information of AppStore* which does not exist.
      required:
        - author
        - category
        - container
        - description
        - developer
        - icon
        - screenshot_link
        - tagline
        - thumbnail
        - title
        - tips
      properties:
        title:
          type: object
          additionalProperties:
            type: string
          example:
            en_US: Syncthing
        description:
          type: object
          additionalProperties:
            type: string
          example:
            en_US: Syncthing is a continuous file synchronization program. It synchronizes files between two or more computers in real time, safely protected from prying eyes. Your data is your data alone and you deserve to choose where it is stored, whether it is shared with some third party, and how it's transmitted over the internet.
        tagline:
          type: object
          additionalProperties:
            type: string
          example:
            en_US: Free, secure, and distributed file synchronisation tool.
        icon:
          type: string
          example: https://cdn.jsdelivr.net/gh/IceWhaleTech/CasaOS-AppStore@main/Apps/Syncthing/icon.png
        screenshot_link:
          type: array
          items:
            type: string
          example:
            - https://cdn.jsdelivr.net/gh/IceWhaleTech/CasaOS-AppStore@main/Apps/OpenSpeedTest/screenshot-1.png
            - https://cdn.jsdelivr.net/gh/IceWhaleTech/CasaOS-AppStore@main/Apps/OpenSpeedTest/screenshot-2.png
          x-oapi-codegen-extra-tags:
            mapstructure: "screenshot_link"
            yaml: "screenshot_link,omitempty"
        thumbnail:
          type: string
          example: https://cdn.jsdelivr.net/gh/IceWhaleTech/CasaOS-AppStore@main/Apps/Jellyfin/thumbnail.jpg
        author:
          type: string
          example: CasaOS Team
        developer:
          type: string
          example: Syncthing
        category:
          type: string
          example: Backup
        tips:
          $ref: "#/components/schemas/TipsStoreInfo"
        container:
          $ref: "#/components/schemas/ContainerStoreInfo"
    ComposeAppContainers:
      description: |-
        Schema for `ComposeAppContainersOK` response.
      properties:
        main:
          type: string
          example: "syncthing"
        containers:
          type: object
          additionalProperties:
            $ref: "#/components/schemas/ContainerSummary"
          example:
            "09e70f4a7dd3fc43de6a1bc40d45c79f857544e13679e5b1a47409034a7ce621":
              {}
    ContainerSummary:
      type: object
      description: |-
        See [Go Doc](https://pkg.go.dev/github.com/docker/compose/v2@v2.14.2/pkg/api#ContainerSummary) for the schema structure of `ContainerSummary`.

      x-go-type: api.ContainerSummary
      x-go-type-import:
        name: api
        path: github.com/docker/compose/v2/pkg/api
    ContainerStoreInfo:
      required:
        - envs
        - index
        - port_map
        - ports
        - volumes
        - devices
      properties:
        port_map:
          type: string
          example: "8384"
          x-oapi-codegen-extra-tags:
            mapstructure: "port_map"
            yaml: "port_map,omitempty"
        index:
          type: string
          example: "/"
          x-oapi-codegen-extra-tags:
            yaml: ",omitempty"
        envs:
          type: array
          items:
            $ref: "#/components/schemas/EnvStoreInfo"
          x-oapi-codegen-extra-tags:
            yaml: ",omitempty"
        ports:
          type: array
          items:
            $ref: "#/components/schemas/PortStoreInfo"
          x-oapi-codegen-extra-tags:
            yaml: ",omitempty"
        volumes:
          type: array
          items:
            $ref: "#/components/schemas/VolumeStoreInfo"
          x-oapi-codegen-extra-tags:
            yaml: ",omitempty"
        devices:
          type: array
          items:
            $ref: "#/components/schemas/DeviceStoreInfo"
          x-oapi-codegen-extra-tags:
            yaml: ",omitempty"
        shell:
          type: string
          example: "/bin/sh"
          x-oapi-codegen-extra-tags:
            yaml: ",omitempty"
    EnvStoreInfo:
      required:
        - container
        - description
        - configurable
      properties:
        container:
          type: string
          example: TZ
        description:
          type: object
          additionalProperties:
            type: string
          example:
            en_US: Time Zone
        configurable:
          $ref: "#/components/schemas/Configurable"
    PortStoreInfo:
      required:
        - container
        - protocol
        - description
        - configurable
      properties:
        container:
          type: string
          example: 8384
        protocol:
          type: string
          enum:
            - tcp
            - udp
        description:
          type: object
          additionalProperties:
            type: string
          example:
            en_US: Web UI
        configurable:
          $ref: "#/components/schemas/Configurable"
    VolumeStoreInfo:
      required:
        - container
        - description
        - configurable
      properties:
        container:
          type: string
          example: /config
        description:
          type: object
          additionalProperties:
            type: string
          example:
            en_US: Syncthing config directory
        configurable:
          $ref: "#/components/schemas/Configurable"
    DeviceStoreInfo:
      properties:
        container:
          type: string
          example: /dev/dri
        description:
          type: object
          additionalProperties:
            type: string
          example:
            en_US: GPU device
        configurable:
          $ref: "#/components/schemas/Configurable"
    TipsStoreInfo:
      required:
        - before_install
      properties:
        before_install:
          type: array
          items:
            $ref: "#/components/schemas/Tip"
          x-oapi-codegen-extra-tags:
            mapstructure: "before_install"
            yaml: "before_install,omitempty"
    Configurable:
      type: string
      enum:
        - yes
        - no
        - basic
        - advanced
    Tip:
      required:
        - content
      properties:
        content:
          type: object
          additionalProperties:
            type: string
          example:
            en_US: "Username: "
        value:
          type: string
          example: "admin"
          x-oapi-codegen-extra-tags:
            yaml: ",omitempty"